Don't specify the icon name with extension. Closes: #551544.
According to the [freedesktop.org standard][0], there is no need to
include the extension of the name of the icons if no absolute pathname
is used.

Therefore, not keeping it makes things more flexible if we decide to
reimplement the icons in other formats in the future.

This, BTW, fixes Debian's bug #551544.

[0]: http://standards.freedesktop.org/desktop-entry-spec/1.0/ar01s05.html